Improve SQL tests
Separate the tests for Squirrel and database/sql. Add tests for
database/sql/driver.
